Audio changes
alsa: support non-mmap
audio intercept: fix ringbuffer implementation
don't let audio lag because of video: closes #34

+++ capsulerun/src/encoder.h
@@ -44,7 +44,7 @@ struct AudioFormat {
 };
 
 typedef int (*VideoFormatReceiver)(void *private_data, VideoFormat *vfmt);
-typedef int (*VideoFrameReceiver)(void *private_data, uint8_t *buffer, size_t buffer_size, int64_t *timestamp);
+typedef int64_t (*VideoFrameReceiver)(void *private_data, uint8_t *buffer, size_t buffer_size, int64_t *timestamp);
 
 typedef int (*AudioFormatReceiver)(void *private_data, AudioFormat *afmt);
 typedef void* (*AudioFramesReceiver)(void *private_data, int64_t *num_frames);
